LayoutTests/fast/dom/shadow/shadow-nested-pseudo-id-expected.txt

Nested pseudo id should be matchable

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(tuple.shadowChild).color is 'rgb(255, 0, 0)'
PASS window.getComputedStyle(shadowTuple.shadowChild).color is 'rgb(0, 0, 255)'
PASS window.getComputedStyle(shadowTuple.shadowChild).backgroundColor is 'rgb(0, 128, 0)'
PASS successfullyParsed is true

TEST COMPLETE
